python-grpc-server/pyproject.toml
[project]
name = "python-grpc-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
]

[tool.black]
exclude = '''
    (
        # Ignore virtual environment files
        \.venv
        # Ignore files generated by grpc
        | .*_pb2\.py$
        | .*_pb2.pyi$
        | .*_pb2_grpc.py$
    )
'''
